Multithreading working, but not finished
Multithreading works, but still needs documenting and revising. There
are multiple places that can be investigated for speedup:

- Check whether to always to submit events/activities to the thread_handler or whether to first check inside own thread.

- Perhaps always submitting inside thread and letting thread_handler
only still activites/events for load balancing is better.

- Change the interval time for distributing and stealing activites/events from the threads (this locks the queue).

- Some elements can perhaps be removed from the shared
Arc<Mutex<..>> structs, for them to not be locked by
threads/thread_handler when submitting/removing elements. For example,
the InnerConstellation struct is pretty big.

- And more..